+++ client/src/app/(withauth)/products/fabric/database/page.tsx
@@ -1,7 +1,7 @@
 "use client"
 import Error from "@/components/error"
 import Loading from "@/components/loading"
-import Show from "@/components/show"
+import Suspense from "@/components/suspense"
 import { endPoints } from "@/constants/api-endpoints"
 import HTTPMethods from "@/constants/http-methods"
 import useConfirm from "@/hooks/use-confirm"
@@ -48,39 +48,31 @@ export default function Page() {
   }
 
   return (
-    <Fragment>
-      <Show when={!db?.isLoading}>
-        <Show when={!!dbId && !db.error}>
-          <Container>
-            <Hero>
-              <p className="branding text-capitalize">{db?.data?.db?.name}</p>
-              <p className="muted-text mt-3">Your Db kvs will be displayed below (if any)</p>
-              <SensitiveInfoPanel credentialIcon={<CubeIcon />} credentialName="DB ID" credentialValue={db?.data?.db?.dbId} />
-              <SensitiveInfoPanel credentialIcon={<LockOpen2Icon />} credentialName="DB Password" credentialValue={db?.data?.db?.dbPassword} />
-              <Button variant="danger" onClick={deleteDb}>Delete Database<TrashIcon className="icon-right" /></Button>
-            </Hero>
-            <Show when={!!db?.data?.kvs && db?.data?.kvs.length}>
-              <h4 className="text-white">KVs</h4>
-              <Table responsive hover variant="light">
-                <thead>
-                  <tr>
-                    <th>Key</th>
-                    <th>Value</th>
-                  </tr>
-                </thead>
-                {displayKvs()}
-              </Table>
-            </Show>
-            {confirmDialog()}
-          </Container>
-        </Show>
-        <Show when={!dbId || !!db.error}>
-          <Error />
-        </Show>
-      </Show>
-      <Show when={db?.isLoading}>
-        <Loading />
-      </Show>
-    </Fragment >
+    <Suspense condition={!db?.isLoading} fallback={<Loading />}>
+      <Suspense condition={!!dbId && !db.error} fallback={<Error />}>
+        <Container>
+          <Hero>
+            <p className="branding text-capitalize">{db?.data?.db?.name}</p>
+            <p className="muted-text mt-3">Your Db kvs will be displayed below (if any)</p>
+            <SensitiveInfoPanel credentialIcon={<CubeIcon />} credentialName="DB ID" credentialValue={db?.data?.db?.dbId} />
+            <SensitiveInfoPanel credentialIcon={<LockOpen2Icon />} credentialName="DB Password" credentialValue={db?.data?.db?.dbPassword} />
+            <Button variant="danger" onClick={deleteDb}>Delete Database<TrashIcon className="icon-right" /></Button>
+          </Hero>
+          <Suspense condition={!!db?.data?.kvs && db?.data?.kvs.length} fallback={null}>
+            <h4 className="text-white">KVs</h4>
+            <Table responsive hover variant="light">
+              <thead>
+                <tr>
+                  <th>Key</th>
+                  <th>Value</th>
+                </tr>
+              </thead>
+              {displayKvs()}
+            </Table>
+          </Suspense>
+          {confirmDialog()}
+        </Container>
+      </Suspense>
+    </Suspense>
   )
 }