+++ client/src/app/(withauth)/subscription/page.tsx
@@ -1,7 +1,7 @@
 "use client"
 import { Fragment, useContext, useState } from "react"
 import { GlobalContext } from "@/context/globalstate.provider"
-import Show from "@/components/show"
+import Suspense from "@/components/suspense"
 import { toast } from "react-hot-toast"
 import { endPoints } from "@/constants/api-endpoints"
 import HTTPMethods from "@/constants/http-methods"
@@ -90,7 +90,7 @@ export default function Page() {
 
   return (
     <Fragment>
-      <Show when={!pricingDetails.isLoading}>
+      <Suspense condition={!pricingDetails.isLoading} fallback={<Loading />}>
         <div className="box">
           <p className="branding">Subscribe & Usage</p>
           <p className="muted-text mb-4">Subscribe & Track your API Credentials usage from here</p>
@@ -99,22 +99,19 @@ export default function Page() {
           <InfoPanel infoIcon={<CalendarIcon />} infoName="Validity" infoValue={userState.hasActiveSubscription ? `Valid upto ${moment(userState.expiresAt).format("MMM, Do YYYY")}` : "No Validity Data"} />
           <InfoPanel infoIcon={<PieChartIcon />} infoName="Subscription Usage" infoValue={userState.hasActiveSubscription ? `${userState.remainingCredits} / ${pricingDetails.data?.[`${userState.selectedPlan.toLowerCase()}`]?.grantedCredits} Credits remaining` : "No Subscriptions Usage Data"} />
           <Fragment>
-            <Show when={displayTrialButton && !userState.hasActiveSubscription}>
+            <Suspense condition={displayTrialButton && !userState.hasActiveSubscription} fallback={null}>
               <Button variant="primary" className="btn-block" onClick={activateTrial}>Activate Trial<ArrowRightIcon className="icon-right" /></Button>
-            </Show>
-            <Show when={!userState.hasActiveSubscription}>
+            </Suspense>
+            <Suspense condition={!userState.hasActiveSubscription} fallback={null}>
               <Button variant="primary" className="btn-block" type="submit" disabled={isTxProcessing} onClick={activatePro}>
-                <Show when={!isTxProcessing}>Activate Pro {pricingDetails.data?.pro?.price} MATIC<ArrowRightIcon className="icon-right" /></Show>
-                <Show when={isTxProcessing}><i className="fas fa-circle-notch fa-spin"></i> Processing Payment</Show>
+                <Suspense condition={!isTxProcessing} fallback={null}>Activate Pro {pricingDetails.data?.pro?.price} MATIC<ArrowRightIcon className="icon-right" /></Suspense>
+                <Suspense condition={isTxProcessing} fallback={null}><i className="fas fa-circle-notch fa-spin"></i> Processing Payment</Suspense>
               </Button>
-            </Show>
+            </Suspense>
           </Fragment>
         </div>
-      </Show>
-      <Show when={pricingDetails.isLoading}>
-        <Loading />
-      </Show>
+      </Suspense>
       {confirmDialog()}
-    </Fragment >
+    </Fragment>
   )
 }