+++ client/src/components/identity-guard.tsx
@@ -3,7 +3,7 @@ import { Fragment, useState } from "react"
 import axios from "axios"
 import { Button, Form } from "react-bootstrap"
 import { uiConstants } from "@/constants/global-constants"
-import Show from "@/components/show"
+import Suspense from "@/components/suspense"
 import { endPoints } from "@/constants/api-endpoints"
 import { toast } from "react-hot-toast"
 import { ArrowRightIcon } from "@radix-ui/react-icons"
@@ -62,7 +62,7 @@ export default function IdentityGuard({ onIdentitySuccess, onIdentityFailure }: 
 
   return (
     <Fragment>
-      <Show when={identityStep === 1}>
+      <Suspense condition={identityStep === 1} fallback={null}>
         <form className="box" onSubmit={generatePassKey}>
           <p className="branding">Identity</p>
           <p className="muted-text">Enter the email address to get started</p>
@@ -71,13 +71,14 @@ export default function IdentityGuard({ onIdentitySuccess, onIdentityFailure }: 
             <Form.Control disabled={isLoading} autoFocus type="email" placeholder="someone@example.com" onChange={(e) => setState({ ...state, email: e.target.value })} required autoComplete={"off"} minLength={4} maxLength={40} />
           </Form.Group>
           <Button variant="primary" type="submit" disabled={isLoading} className="mt-1 btn-block">
-            <Show when={!isLoading}>Get Identity Passkey <ArrowRightIcon className="icon-right" /></Show>
-            <Show when={isLoading}><i className="fas fa-circle-notch fa-spin"></i> {alert}</Show>
+            <Suspense condition={!isLoading} fallback={<><i className="fas fa-circle-notch fa-spin"></i> {alert}</>}>
+              Get Identity Passkey <ArrowRightIcon className="icon-right" />
+            </Suspense>
           </Button>
           <p className="muted-text mt-1">By using {uiConstants.brandName}, you agree to our Terms of Service and Privacy Policy.</p>
         </form>
-      </Show>
-      <Show when={identityStep === 2}>
+      </Suspense>
+      <Suspense condition={identityStep === 2} fallback={null}>
         <form className="box" onSubmit={verifyPassKey}>
           <p className="branding">Identity</p>
           <p className="muted-text">Please verify your identity by entering the identity passkey we sent to your inbox.</p>
@@ -86,12 +87,13 @@ export default function IdentityGuard({ onIdentitySuccess, onIdentityFailure }: 
             <Form.Control type="password" disabled={isLoading} name="passKey" placeholder="XXXX-XXXX" onChange={(e) => setState({ ...state, passKey: e.target.value })} required autoComplete={"off"} />
           </Form.Group>
           <Button variant="primary" type="submit" disabled={isLoading} className="mt-4 btn-block">
-            <Show when={!isLoading}>Continue <ArrowRightIcon className="icon-right" /></Show>
-            <Show when={isLoading}><i className="fas fa-circle-notch fa-spin"></i> {alert}</Show>
+            <Suspense condition={!isLoading} fallback={<><i className="fas fa-circle-notch fa-spin"></i> {alert}</>}>
+              Continue <ArrowRightIcon className="icon-right" />
+            </Suspense>
           </Button>
           <p id="alert" className="mt-1 mb-1">{identityAlert}</p>
         </form>
-      </Show>
+      </Suspense>
     </Fragment >
   )
 }